api_gateway/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client for Convex integration
httpx==0.27.2
websockets==13.1

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx-sse==0.4.0

# Development tools
black==24.10.0
isort==5.13.2
mypy==1.13.0
pylint==3.3.2

# Additional utilities
python-multipart==0.0.12
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4